project up  ch18 exercise 27 update

+++ src/ch18/ex27/codeb/CodeB.java
@@ -14,24 +14,24 @@ public class CodeB {
         System.out.println("\n====CODE_B===");
         System.out.println("\nGZIP Compress Check\n");
 
+        // remove to run programm   блокировка чтобы не писало файлы
+        int k = 0;
+        if (k == 0) {
+            return;
+        }
+        // remove to run programm   блокировка чтобы не писало файлы
+
         GZIPCompress.check(args);  // байтовый, символьный, строчный ввод, вывод работают
         ZipCompress.check(args);   // байтовый, символьный, строчный ввод, вывод работают
-
-
-        File[] files = Catalog.local("./src/ch18/ex27/codeb/archives",".*");
+        File[] files = Catalog.local("./src/ch18/ex27/codeb/archives", ".*");
         for (File file : files) {
             file.delete();
         }
-        files = Catalog.local("./src/ch18/ex27/codeb/",".*.[gz][iz].*");
+        files = Catalog.local("./src/ch18/ex27/codeb/", ".*.[gz][iz].*");
         for (File file : files) {
             System.out.println(file.getName());
-            file.renameTo(new File(file.getParent()+"\\archives\\"+file.getName()));
+            file.renameTo(new File(file.getParent() + "\\archives\\" + file.getName()));
         }
-
-
-
-
-
 
 
     }
